Profile Id data type changed from Int -> Long through Room Migration

+++ app/src/main/java/com/example/foundit/presentation/screens/profile/components/ProfileHeadingCard.kt
@@ -49,7 +49,7 @@ fun ProfileHeadingCard(
     firstName: String,
     lastName: String,
     profileCountryCode: Int,
-    profileId: Int,
+    profileId: Long,
     onEditProfileClick: () -> Unit,
 ) {
     Card(
